requirements.txt
joblib==1.4.2
pandas==2.2.3
pydantic_settings==2.7.1
python-dotenv==1.0.1
scapy==2.6.1
scikit_learn==1.6.1
SQLAlchemy==2.0.38
websockets==15.0
